Use uwsgi between Apache and Keystone
This switches from using http to uwsgi between Apache and Keystone.
If the uwsgi port setting is used, then mod_proxy_uwsgi will be used.
Otherwise if the file socket setting is used, then the uwsgi module will be used.
Both Ubuntu and Redhat will be able to use the port setting, but
only Ubuntu can use the file socket setting. The file socket was
previously broken but now will work on Ubuntu with this change.